@moberegger 2025-05-29
> Which do you think is the better approach overall?

I think this draft is more likely to be merged upstream compared to #4, so my gut says this one is the better option.