@Insomniak47 2025-05-30
> This is definitely true for us, and I imagine that's the case for most people. But who knows what is happening in the wild, so I can see the other approach being held under more scrutiny.

I think that most libraries in some way or another favour these performance characteristics anyways since it is such a common pattern (things that cache key execs, things with formatting optimizations etc). I think that you're safe either way but I'm happy to go with the one more likely to be upstreamed